[refactor] User의 기능 구현
User 게임 참가자의 공통된 기능 구현
Player User 리팩토링에 따른 구조 변경
Dealer User 리팩토링에 따른 구조 변경
