+++ src/main/java/domain/user/Dealer.java
@@ -1,37 +1,18 @@
 package domain.user;
 
-import domain.HandCard;
-import domain.card.Card;
-
-public class Dealer {
-    public static final String DEALER_NAME = "딜러";
-    private final HandCard handCard;
+public class Dealer extends User{
+    private static final String DEALER_NAME = "딜러";
+    public static final int DRAW_MAX_SCORE = 16;
 
     public Dealer() {
-        handCard = new HandCard();
-    }
-
-    public String getName() {
-        return DEALER_NAME;
-    }
-
-    public void draw(Card card) {
-        handCard.add(card);
-    }
-
-    public String getStatus() {
-        StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append(DEALER_NAME)
-                .append(": ")
-                .append(handCard.getNames());
-        return stringBuilder.toString();
+        super(DEALER_NAME);
     }
 
     public String getFirstStatus(){
         return getStatus().split(",")[0];
     }
 
-    public int getScore() {
-        return handCard.getScore();
+    public boolean isUnderBound() {
+        return getScore() <= DRAW_MAX_SCORE;
     }
 }
